Word clouds
Here in this code file, we are going to visualize the terms that are more frequent using word clouds. We are going to build three word clouds, one for entire corpus and one each for two classes based on target class. We are also going to show most frequent terms using barplot also. 

Words that appear frequent in both target classes serve less for classification. So, we are going to save them into an object and delete them from the data in further analysis.